pyproject.toml
[tool.poetry]
name = "govbot-pkg"
version = "0.1.0"
description = "A twitter bot that tweets about crypto governance proposals"
authors = ["stephenthoma <stephen@thoma.io>"]
readme = "README.md"
packages = [{include = "govbot"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
tweepy = "^4.12.1"
google-cloud-firestore = "^2.9.0"
sgqlc = "^16.1"
pytz = "^2022.7.1"
google-cloud-tasks = "^2.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
